+++ src/Checkout/resources/config/module.php
@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 return [
+    'event_listeners' => true,
     'store' => [
         'driver' => 'request'
     ]